First implementation of general ML class
